Change index type to size_t

+++ yrt-pet/include/yrt-pet/datastruct/projection/ProjectionData.hpp
@@ -74,7 +74,7 @@ public:
 	virtual void
 	    getProjectionProperties(ProjectionProperties& props,
 	                            const ProjectionPropertyManager& propManager,
-	                            bin_t bin, int tid) const;
+	                            bin_t bin, size_t tid) const;
 
 	Line3D getLOR(bin_t bin) const;
 	virtual void clearProjections(float value);
